Remove dead MessageList.ts and deduplicate message rendering logic
Problem: ChatPanel.ts duplicated ~280 lines of message rendering logic
from MessageList.ts (StableUserMessage, StableAssistantMessage,
renderUserMessage, renderAssistantMessage, countResolvedToolResults,
isCollapsibleUserMessage). MessageList.ts was left as dead code after
being superseded by ChatPanel.ts in the dockview workspace refactor.

Fix: Move the shared rendering functions to message-renderers.ts (which
already serves as the shared rendering module for ChatPanel and
SubagentView). Update ChatPanel.ts to import these functions from
message-renderers.ts. Delete MessageList.ts since it is no longer
imported anywhere. Also update the ChatPanel.ts file comment to
accurately describe what it does.
